Refactor file descriptors as stated in the code review
