+++ dbus-java-transport-junixsocket/src/main/java/module-info.java
@@ -10,7 +10,4 @@ module org.freedesktop.dbus.transport.junixsocket {
 
     provides org.freedesktop.dbus.spi.message.ISocketProvider
             with org.freedesktop.dbus.transport.junixsocket.JUnixSocketSocketProvider;
-
-    provides org.freedesktop.dbus.spi.filedescriptors.IFileDescriptorHelper
-            with org.freedesktop.dbus.transport.junixsocket.JUnixSocketFileDescriptorHelper;
 }
