--- dbus-java-transport-junixsocket/src/main/resources/META-INF/services/org.freedesktop.dbus.spi.filedescriptors.IFileDescriptorHelper
@@ -1,1 +1,0 @@
-org.freedesktop.dbus.transport.junixsocket.JUnixSocketFileDescriptorHelper
